+++ resharper/resharper-unity/src/CSharp/Feature/Services/Bulbs/ConvertFromCoroutineBulbAction.cs
@@ -24,14 +24,12 @@ namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.Bulbs
             if (element == null) return null;
 
             var unityApi = solution.GetComponent<UnityApi>();
+            var knownTypesCache = solution.GetComponent<KnownTypesCache>();
             var eventFunction = unityApi.GetUnityEventFunction(element);
 
-            IType returnType = TypeFactory.CreateTypeByCLRName(eventFunction.ReturnType, myMethodDeclaration.GetPsiModule());
-            if (eventFunction.ReturnTypeIsArray)
-                returnType = TypeFactory.CreateArrayType(returnType, 1);
+            var returnType = eventFunction.ReturnType.AsIType(knownTypesCache, myMethodDeclaration.GetPsiModule());
 
-            var language = myMethodDeclaration.Language;
-            var changeTypeHelper = LanguageManager.Instance.GetService<IChangeTypeHelper>(language);
+            var changeTypeHelper = LanguageManager.Instance.GetService<IChangeTypeHelper>(myMethodDeclaration.Language);
             changeTypeHelper.ChangeType(returnType, element);
 
             return null;