+++ resharper/resharper-unity/src/UnityEventFunction.cs
@@ -7,6 +7,7 @@ using JetBrains.ReSharper.Psi;
 using JetBrains.ReSharper.Psi.CSharp;
 using JetBrains.ReSharper.Psi.CSharp.Impl;
 using JetBrains.ReSharper.Psi.CSharp.Tree;
+using JetBrains.ReSharper.Psi.Modules;
 using JetBrains.Util;
 
 namespace JetBrains.ReSharper.Plugins.Unity
@@ -17,8 +18,8 @@ namespace JetBrains.ReSharper.Plugins.Unity
         private readonly Version myMaximumVersion;
 
         public UnityEventFunction([NotNull] string name, [NotNull] IClrTypeName typeName,
-                                  [NotNull] IClrTypeName returnType, bool returnTypeIsArray, bool isStatic,
-                                  bool canBeCoroutine, string description, bool undocumented, Version minimumVersion,
+                                  [NotNull] UnityTypeSpec returnType, bool isStatic, bool canBeCoroutine,
+                                  string description, bool undocumented, Version minimumVersion,
                                   Version maximumVersion, [NotNull] params UnityEventFunctionParameter[] parameters)
         {
             Description = description;
@@ -30,15 +31,13 @@ namespace JetBrains.ReSharper.Plugins.Unity
             Name = name;
             TypeName = typeName;
             ReturnType = returnType;
-            ReturnTypeIsArray = returnTypeIsArray;
             Parameters = parameters.Length > 0 ? parameters : EmptyArray<UnityEventFunctionParameter>.Instance;
         }
 
         [NotNull] public IClrTypeName TypeName { get; }
         [NotNull] public string Name { get; }
         [NotNull] public UnityEventFunctionParameter[] Parameters { get; }
-        [NotNull] public IClrTypeName ReturnType { get; }
-        public bool ReturnTypeIsArray { get; }
+        [NotNull] public UnityTypeSpec ReturnType { get; }
         public bool CanBeCoroutine { get; }
         public bool IsStatic { get; }
         [CanBeNull] public string Description { get; }
@@ -46,12 +45,17 @@ namespace JetBrains.ReSharper.Plugins.Unity
 
         [NotNull]
         public IMethodDeclaration CreateDeclaration([NotNull] CSharpElementFactory factory,
+                                                    [NotNull] KnownTypesCache knownTypesCache,
                                                     [NotNull] IClassLikeDeclaration classDeclaration,
                                                     AccessRights accessRights,
                                                     bool makeVirtual = false,
                                                     bool makeCoroutine = false)
         {
             var builder = new StringBuilder(128);
+            var args = new object[1 + Parameters.Length];
+            object arg;
+            var argIndex = 0;
+            var module = classDeclaration.GetPsiModule();
 
             if (accessRights != AccessRights.NONE)
             {
@@ -63,13 +67,21 @@ namespace JetBrains.ReSharper.Plugins.Unity
 
             // Consider this declaration a template, and the final generated code implements (or overrides) this API
             if (makeVirtual) builder.Append("virtual ");
-            builder.Append("global::");
             if (makeCoroutine && CanBeCoroutine)
                 builder.Append(PredefinedType.IENUMERATOR_FQN.FullName);
             else
             {
-                builder.Append(ReturnType.FullName);
-                if (ReturnTypeIsArray) builder.Append("[]");
+                arg = GetTypeObject(ReturnType, knownTypesCache, module);
+                if (arg is string)
+                {
+                    builder.Append(arg);
+                    if (ReturnType.IsArray) builder.Append("[]");
+                }
+                else
+                {
+                    builder.Append("$0");
+                    args[argIndex++] = arg;
+                }
             }
 
             builder.Append(" ");
@@ -85,16 +97,25 @@ namespace JetBrains.ReSharper.Plugins.Unity
                 // From reflection point of view, it's a "ByRef" Type, and that's all we know...
                 // The only place it's currently being used is an out parameter
                 if (parameter.IsByRef) builder.Append("out ");
-                builder.Append("global::");
-                builder.Append(parameter.ClrTypeName.FullName);
-                if (parameter.IsArray) builder.Append("[]");
+
+                arg = GetTypeObject(parameter.TypeSpec, knownTypesCache, module);
+                if (arg is string)
+                {
+                    builder.Append(arg);
+                    if (parameter.TypeSpec.IsArray) builder.Append("[]");
+                }
+                else
+                {
+                    builder.Append($"${argIndex}");
+                    args[argIndex++] = arg;
+                }
                 builder.Append(' ');
                 builder.Append(parameter.Name);
             }
 
             builder.Append(");");
 
-            var declaration = (IMethodDeclaration) factory.CreateTypeMemberDeclaration(builder.ToString());
+            var declaration = (IMethodDeclaration) factory.CreateTypeMemberDeclaration(builder.ToString(), args);
             declaration.SetResolveContextForSandBox(classDeclaration, SandBoxContextType.Child);
             return declaration;
         }
@@ -108,6 +129,18 @@ namespace JetBrains.ReSharper.Plugins.Unity
         public bool SupportsVersion(Version unityVersion)
         {
             return myMinimumVersion <= unityVersion && unityVersion <= myMaximumVersion;
+        }
+
+        private static object GetTypeObject(UnityTypeSpec typeSpec, KnownTypesCache knownTypesCache, IPsiModule module)
+        {
+            if (typeSpec.TypeParameters.Length == 0)
+            {
+                var keyword = CSharpTypeFactory.GetTypeKeyword(typeSpec.ClrTypeName);
+                if (keyword != null)
+                    return keyword;
+            }
+
+            return typeSpec.AsIType(knownTypesCache, module);
         }
     }
 }