+++ resharper/resharper-unity/src/Yaml/Feature/Services/Navigation/UnityScriptsOccurrence.cs
@@ -1,10 +1,8 @@
 using System;
-using JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.DeferredCaches.AssetHierarchy.Elements;
 using JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.DeferredCaches.AssetHierarchy.References;
 using JetBrains.ReSharper.Psi;
 using JetBrains.ReSharper.Psi.Pointers;
 using JetBrains.UI.RichText;
-using JetBrains.Util.Extension;
 
 namespace JetBrains.ReSharper.Plugins.Unity.Yaml.Feature.Services.Navigation
 {
@@ -24,10 +22,10 @@ namespace JetBrains.ReSharper.Plugins.Unity.Yaml.Feature.Services.Navigation
             var declaredElement = DeclaredElementPointer.FindDeclaredElement();
             if (declaredElement == null)
                 return base.GetDisplayText();
-            
-            if (UnityApi.IsDescendantOfScriptableObject(declaredElement as IClass))
-                return SourceFile.GetLocation().Name;  
-            
+
+            if ((declaredElement as IClass).DerivesFromScriptableObject())
+                return SourceFile.GetLocation().Name;
+
             return base.GetDisplayText();
         }
 
@@ -37,9 +35,12 @@ namespace JetBrains.ReSharper.Plugins.Unity.Yaml.Feature.Services.Navigation
                 return null;
             return base.GetRelatedFilePresentation();
         }
-        
-        private bool IsRelatedToScriptableObject() => UnityApi.IsDescendantOfScriptableObject(DeclaredElementPointer.FindDeclaredElement() as IClass);
-        
+
+        private bool IsRelatedToScriptableObject()
+        {
+            return (DeclaredElementPointer.FindDeclaredElement() as IClass).DerivesFromScriptableObject();
+        }
+
         public override string ToString()
         {
             return $"Guid: {myGuid:N}";